web-app/requirements.txt
Flask==3.0.0
pydantic==2.5.3
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.11
opensearch-py==2.4.2
haystack-ai
opensearch-haystack
ollama-haystack


